requirements.txt
# General

numpy
matplotlib
pynetworktables  # comms
opencv-python
numba  # Code optimization
tqdm  # nifty progress bar stuff
networkx  # Graph Theory
evdev  # Controller
ikpy

# Hardware
adafruit-circuitpython-servokit  # Arm
rplidar-roboticia  # Lidar
pyserial  # Serial communication. Used for Rosmaster

# these only apply if not using rosmaster.
#board  # Communicating with INA219
#smbus
#adafruit-circuitpython-ina219
#imusensor


# other
noise
plotly
torch
tensorrt
filterpy  # for sort tracking
lap  # linear assignment for sort

# Centerline Calculation
# scikit-image

# Check the installer.sh script